Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Remove IConfig and IOemHook hals.

Change-Id: Id13450143021c28fa751a74877756ce02c2ec78d
CRs-Fixed: 2287149